FIND-291: Replace tnaCard with new searchResults macro
We've established that tnaCard is not a suitable component for the new
design (partly because it should have only one 'supertitle' and partly
because it exists for a different purpose). In response, we have decided
to create a new `search_results` macro for the catalogue that reflects
the design and borrows functionality from tnaCard (such as the ability
to click the whole card).