FIND-291: Search result component (#83)
* Fix: reword typo in class name

* FIND-291: Add clean_description to Record model

The API has been updated to provide a cleanDescription field that

1. removes all XML and HTML tags
2. introduces HTML `<mark>` tags for search term highlighting

The API changes have been described in a Jira ticket comment[^1]

While I'm not a backend developer, I needed to use this field in to
complete FIND-291, so I have replicated what exists for other fields
in the model and in tests. Unfortunately, I can't run the tests locally
because I don't have the necessary config so will rely on CI to verify
they're passing OK.

[^1]: https://national-archives.atlassian.net/browse/FIND-179?focusedCommentId=104429

* FIND-291: Replace tnaCard with new searchResults macro

We've established that tnaCard is not a suitable component for the new
design (partly because it should have only one 'supertitle' and partly
because it exists for a different purpose). In response, we have decided
to create a new `search_results` macro for the catalogue that reflects
the design and borrows functionality from tnaCard (such as the ability
to click the whole card).

---------

Co-authored-by: Andrew Hosgood <andrew.hosgood@nationalarchives.gov.uk>