Refactor macro naming for clarity and consistency
While exploring availability of properties I can use for FIND-277, I was confused by the macro
approach that was in place. This was partly because the inclusion of each macro is mutually exclusive,
but they use to opposite 'if' statements, rather than a more explicit 'if'/'else'. I was also confused
by the way the conditions were included to wrap the entire macro, rather than being used outside the
macro to determine which macro is appropriate.

What I've tried to do here is make the intent clearer by:

* Renaming the files and macros to explain that both are 'information', but one is the information to
  show where the details page is below series level and the other is for details pages at series level
  and above
* Change the condition from two 'if's to an 'if else' that is in the template rather than the macro.
  My goal here is to make their mutual exclusivity immediately apparent

What I haven't done:

* I haven't merged the two macros into one, again because I think they're easier to reason about as two
  separate macros - each for a specific set of levels. My sense here is that it's better to avoid almost
  duplicate HTML within a single macro