FIND-291: Truncate search results while preserving <mark> tags
This commit introduces a new truncate_string_preserving_mark_tags Jinja2
filter. It accepts an input string and a max_length, and returns
a string of the max length. Crucially, it also preserves any <mark> tags
present in the input string

This is a complex function, and I used Copilot to help me formulate it.
I could have worked it out myself but it would have taken me much longer.

Because of this I've:

1. Included a comprehensive suite of unit tests for the functionality
2. Heavily annotated the function with comments, so it will be easier
   for the reviewer to follow