Add content_warning macro and conditional padding
The content warning simply wraps tnaWarning with some
additional HTML structure to control its presentation within
details pages.

A conditional class also been introduced to manage padding of
the container where the content warning is applied. This was
necessary because the design shows less space where the banner
is present.