FIND-394: Add macro for filters not applied banner
Introduces a new macro that will be displayed when the
backend work necessary to show it has been done.

To avoid introducing custom CSS, the HTML uses most of the
classes provided by tnaWarning. It's a little different, though,
because but the Catalogue implementation also needs a visible
heading (which doesn't seem to be supported currently in the
Design System macro[^1]

[^1]: https://design-system.nationalarchives.gov.uk/components/warning/