FIND-407: Temporarily hard-code TNA record count
To address discrepancies between the 'bucket' count and the
count shown at the top-level of the hierarchy, the team has
decided to temporarily hard-code "Over 27 million" as the
value for records held by TNA at the top level of the hierarchy.

Because same macro is used by TNA and non-TNA records, I've
chosen to add a Nunjucks filter which does the replacement
based on the value of `held_by`.

You might feel this commit is a bit regressive and wonder why
I haven't converted the real count to a string representation.
Here's why: the actual number being returned by the backend is
less than 26 million, which I can only assume is part of the
reason backend work is needed to get a more accurate number.