FIND-401: Use new `description.short` API property
The API has been updated to include a new `short` property on `description`.

This new property is the original description with:

1. all HTML tags stripped
2. "..." appended at the end

This commit updates the Record model to include this property, where it
is available, as `short_description` and to include this in the template

While working on this, I've noticed a couple of things

1. the "..." always appears, even if the original description wasn't long
   enough to require truncation
2. we're using three literal "." characters for the ellipsis, rather than
   the HTML entity[^1]

I've described this to the Product Manager and we'll be asking the API
developers to tweak the implementation.

[^1]: https://www.w3.org/wiki/Common_HTML_entities_used_for_typography