FIND-418: Add ellipsis to short description where needed
Introduces the unicode Ellipsis HTML entity where the short description is less
than 250 characters.

Doing this on the frontend is necessary because the API developers are unable to
provide the ellipsis in the API response (they've explained they're use of RegEx
to truncate descriptions means they cannot do this). They have also explained
that their use of word boundaries means the short description could be a bit
longer than 250 characters.

The condition has been introduced because it's possible a short description is
identical to the description (i.e. it wasn't long enough in the first place to
require truncation), in which case an ellipsis is misleading.

I suppose it remains possible a description may have been exactly 250 characters
in which case the ellipsis could still be misleading, but there doesn't seem to
be anything more I can do on the frontend

You might ask why I don't compare the short description to the description in the
Record model and apply the ellipsis if they are different. Again, this isn't an
option becaues the description could contain HTML, which is why we asked for the
API to include a short version.