FIND-418: Truncate description with ellipsis in series box
K-Int have updated the API to introduce a `description.noHtml` property
that we can use to safely truncate the description. This is in place of
the `description.short` property they had initially provided before it
transpired the technical limitations of Rosetta meant it could not do
what we needed.

I have therefore effectively replaced `short_description` with a new
`no_html_description` property which is truncated using Jinja2 on the
front end.