Fix: Hide 'How to order' accordion item where it would be empty
Bug spotted in production today (see Slack thread) where the accordion
title and content are collapsed. This seems to be because the backend
isn't providing the strings we'd need. That will need to be looked into
by a BE developer, but this quick fix will at least stop the accordion
having an empty item with a collapsed heading.